refactor: move popular data source to subpackage
